requirements.txt
pandas
numpy
ipykernel
pyyaml
from_root
-e .